PROG-234 added some stuff to test
