+++ src/main/java/com/lightningrobotics/voidrobot/commands/auto/commands/AutonIndexeCargo.java
@@ -1,0 +1,36 @@
+package com.lightningrobotics.voidrobot.commands.auto.commands;
+
+import com.lightningrobotics.voidrobot.constants.Constants;
+import com.lightningrobotics.voidrobot.subsystems.Indexer;
+
+import edu.wpi.first.wpilibj2.command.CommandBase;
+
+public class AutonIndexeCargo extends CommandBase {
+
+private final Indexer indexer;
+
+    public AutonIndexeCargo(Indexer indexer) {
+        this.indexer = indexer;
+
+        addRequirements(indexer);
+
+    }
+
+    @Override
+    public void initialize() {}
+
+    @Override
+    public void execute() {
+        indexer.setPower(0.5d);
+    }
+
+    @Override
+    public void end(boolean interrupted) {
+        indexer.stop();
+    }
+
+    @Override
+    public boolean isFinished() {
+        return indexer.getColorSensorOutputs() != 0;
+    }
+}
