+++ src/main/java/com/lightningrobotics/voidrobot/commands/auto/paths/ThreeBallVision.java
@@ -2,11 +2,9 @@ package com.lightningrobotics.voidrobot.commands.auto.paths;
 
 import com.lightningrobotics.common.auto.Path;
 import com.lightningrobotics.common.command.core.TimedCommand;
-import com.lightningrobotics.voidrobot.commands.auto.commands.AutonAutoIndex;
-import com.lightningrobotics.voidrobot.commands.auto.commands.AutonAutoIndex2;
+import com.lightningrobotics.voidrobot.commands.auto.commands.AutonIndexeCargo;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonIntake;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonShootCargoVision;
-import com.lightningrobotics.voidrobot.commands.turret.AimTurret;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonShootCargo;
 import com.lightningrobotics.voidrobot.subsystems.*;
 
@@ -27,20 +25,27 @@ public class ThreeBallVision extends ParallelCommandGroup {
 
 			new SequentialCommandGroup(
 
-				new ParallelCommandGroup(
+				// shoots the first ball
+				new ParallelDeadlineGroup(
 					start3Ball.getCommand(drivetrain),
-					new AutonShootCargo(shooter, hood, indexer, turret, 3700d, 0d, 10d)
+					new AutonShootCargo(shooter, hood, indexer, turret, 3700d, 0d, 10d),
+					new AutonIndexeCargo(indexer)
 				),
 
 				new InstantCommand(() -> System.out.println("about to shoot ball two ----------------------------------------")),
 
 				new AutonShootCargoVision(shooter, hood, indexer, turret, vision),
 
-				end3Ball.getCommand(drivetrain), 
+				new ParallelDeadlineGroup(
+					end3Ball.getCommand(drivetrain), 
+					new AutonIndexeCargo(indexer)
+				),
 				
 				new InstantCommand(() -> System.out.println("about to shoot ball three ----------------------------------------")),
 
-				new AutonShootCargoVision(shooter, hood, indexer, turret, vision)
+				new AutonShootCargoVision(shooter, hood, indexer, turret, vision),
+
+				new InstantCommand(() -> System.out.println("we have ended ----------------------------------------------------")) // This line was written by Enoch 
 			)
 		);
